Fix percona-extrabackup.sh for better monitoring
Removes unused "email team if it fails" fucntionality which didn't work with 'set -e' anyway;
Installs file using copy instead of template, which was all templated in a "raw" block;
Removes unneeded "mailutils" dependency

+++ roles/percona-backup/files/percona-xtrabackup.sh
@@ -1,4 +1,3 @@
-{% raw %}
 #!/bin/bash
 #
 # File: percona-xtrabackup.sh
@@ -15,19 +14,25 @@
 #
 # Author: mpatterson@bluebox.net
 
-set -o errexit
 
-email=team-infrastructure@bluebox.net
 backup_script=/usr/bin/innobackupex
 gzip=/bin/gzip
 
 backup_retention_days=7
 
 backup_root_dir=/backup/percona/
+logfile=$backup_root_dir/percona-backup.last.log
 
 # create a new db archive, use tar stream to compress on the fly
-"$backup_script" --user=root --socket=/var/run/mysqld/mysqld.sock --stream=tar "$backup_root_dir" | "$gzip" - > "$backup_root_dir"`/bin/date +"%Y-%m-%d_%H-%M-%S"`.tar.gz || (/bin/echo "failed to create db archive at: `/bin/date`" | mail "$email" -s "Percona backup failed")
+"$backup_script" --user=root --socket=/var/run/mysqld/mysqld.sock --stream=tar "$backup_root_dir" | "$gzip" - > "$backup_root_dir"`/bin/date +"%Y-%m-%d_%H-%M-%S"`.tar.gz
+# Copy return value of backup
+retval=$?
+# Write last exit status and date to logfile
+echo "${retval} $(date +%s)" > $logfile
+# If the return value is non-zero, exit with last status
+if [ $retval -ne 0 ]; then
+  exit $retval
+fi
 
 echo Removing backups older than "$backup_retention_days" days
 find "$backup_root_dir" -maxdepth 1 -type f -mtime +"$backup_retention_days" -print -delete
-{% endraw %}
